Feature: apply auto register, update, delete clothing bin logic when reported more than 3 times

+++ src/main/java/com/droptheclothes/api/controller/OauthController.java
@@ -14,7 +14,9 @@ import com.droptheclothes.api.model.enums.SignType;
 import com.droptheclothes.api.security.LoginCheck;
 import com.droptheclothes.api.service.AppleAuthenticationService;
 import com.droptheclothes.api.service.OauthService;
+import com.droptheclothes.api.utility.MessageConstants;
 import io.swagger.v3.oas.annotations.Operation;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -95,6 +97,10 @@ public class OauthController {
     @DeleteMapping("/api/oauth2/member")
     @Operation(summary = "회원탈퇴 api", description = "회원탈퇴 api")
     public ApiResponse deleteMember(String authorizationCode) {
+        // TODO: 임시 코드 validation check 삭제해야함
+        if (Objects.isNull(authorizationCode)) {
+            throw new IllegalArgumentException(MessageConstants.WRONG_REQUEST_PARAMETER_MESSAGE);
+        }
         oauthService.deleteMember(authorizationCode);
         return new ApiResponse(ApiResponseHeader.create(ResultCode.SUCCESS), new SingleObject<>(null));
     }
